requirements.txt
django>=6.0,<7
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
channels[daphne]>=4.0
django-filter>=24.0
django-cors-headers>=4.0
croniter>=2.0
requests>=2.31
